pyproject.toml
[project]
name = "mongoconnect"
version = "0.1.0"

requires-python = ">=3.10"
dependencies = [
    "dnspython>=2.8.0",
    "ipykernel>=7.1.0",
    "langchain>=0.3.0,<0.4.0",
    "langchain-community>=0.3.0,<0.4.0",
    "neo4j>=6.0.3",
    "pandas>=2.3.3",
    "plotly>=6.4.0",
    "pydeck>=0.9.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "chroma>=0.2.0",
    "langchain-graph-retriever>=0.8.0",
    "langchain-google-genai",
    "langchain-huggingface",
    "langchain-tools>=0.1.34",
    "langchain-google-vertexai>=2.1.2",
    "torch>=2.5.1",
]
